requirements.txt
# This file contains all the dependencies for this project to run
pillow~=10.0.0
pygame~=2.5.1
